bms/budget_service/requirements.txt
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.3.0
boolean.py==5.0
CacheControl==0.14.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.3
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==25.1
django-pwned-passwords==4.1.0
django-ratelimit==4.1.0
django-rest-passwordreset==1.5.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
et_xmlfile==2.0.0
filelock==3.20.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
license-expression==30.4.4
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.2
openpyxl==3.1.5
packageurl-python==0.17.5
packaging==25.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
pipdeptree==2.30.0
platformdirs==4.5.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pwned-passwords-django==5.2.0
py-serializable==2.1.0
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.2.0
rpds-py==0.24.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
toml==0.10.2
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
whitenoise==6.6.0
